refactor: 안쓰는 코드 삭제 및 컴포넌트 재활용
